LPS-29425 - refactoring SchedulerEngineUtil and adding property for disabling scheduler audits
