+++ portal-impl/src/com/liferay/portal/scheduler/SchedulerEngineHelperImpl.java
@@ -1,0 +1,757 @@
+/**
+ * Copyright (c) 2000-2012 Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.scheduler;
+
+import com.liferay.portal.kernel.audit.AuditMessage;
+import com.liferay.portal.kernel.audit.AuditRouterUtil;
+import com.liferay.portal.kernel.cal.DayAndPosition;
+import com.liferay.portal.kernel.cal.Duration;
+import com.liferay.portal.kernel.cal.Recurrence;
+import com.liferay.portal.kernel.cal.RecurrenceSerializer;
+import com.liferay.portal.kernel.json.JSONFactoryUtil;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.messaging.DestinationNames;
+import com.liferay.portal.kernel.messaging.Message;
+import com.liferay.portal.kernel.scheduler.JobState;
+import com.liferay.portal.kernel.scheduler.SchedulerEngine;
+import com.liferay.portal.kernel.scheduler.SchedulerEngineClusterManager;
+import com.liferay.portal.kernel.scheduler.SchedulerEngineHelper;
+import com.liferay.portal.kernel.scheduler.SchedulerEntry;
+import com.liferay.portal.kernel.scheduler.SchedulerException;
+import com.liferay.portal.kernel.scheduler.SchedulerLifecycle;
+import com.liferay.portal.kernel.scheduler.StorageType;
+import com.liferay.portal.kernel.scheduler.Trigger;
+import com.liferay.portal.kernel.scheduler.TriggerFactoryUtil;
+import com.liferay.portal.kernel.scheduler.TriggerState;
+import com.liferay.portal.kernel.scheduler.messaging.SchedulerResponse;
+import com.liferay.portal.kernel.util.CalendarFactoryUtil;
+import com.liferay.portal.kernel.util.InetAddressUtil;
+import com.liferay.portal.kernel.util.ObjectValuePair;
+import com.liferay.portal.kernel.util.ParamUtil;
+import com.liferay.portal.kernel.util.PortalLifecycle;
+import com.liferay.portal.kernel.util.StringPool;
+import com.liferay.portal.model.CompanyConstants;
+import com.liferay.portal.util.PortalUtil;
+import com.liferay.portal.util.PropsValues;
+
+import java.util.ArrayList;
+import java.util.Calendar;
+import java.util.Date;
+import java.util.List;
+
+import javax.portlet.PortletRequest;
+
+/**
+ * @author Michael C. Han
+ */
+public class SchedulerEngineHelperImpl implements SchedulerEngineHelper {
+
+	public void addJob(
+			Trigger trigger, StorageType storageType, String description,
+			String destinationName, Message message,
+			String messageListenerClassName, String portletId,
+			int exceptionsMaxSize)
+		throws SchedulerException {
+
+		if (message == null) {
+			message = new Message();
+		}
+
+		message.put(
+			SchedulerEngine.MESSAGE_LISTENER_CLASS_NAME,
+			messageListenerClassName);
+		message.put(SchedulerEngine.PORTLET_ID, portletId);
+
+		schedule(
+			trigger, storageType, description, destinationName, message,
+			exceptionsMaxSize);
+	}
+
+	public void addJob(
+			Trigger trigger, StorageType storageType, String description,
+			String destinationName, Object payload,
+			String messageListenerClassName, String portletId,
+			int exceptionsMaxSize)
+		throws SchedulerException {
+
+		Message message = new Message();
+
+		message.setPayload(payload);
+
+		addJob(
+			trigger, storageType, description, destinationName, message,
+			messageListenerClassName, portletId, exceptionsMaxSize);
+	}
+
+	public void addScriptingJob(
+			Trigger trigger, StorageType storageType, String description,
+			String language, String script, int exceptionsMaxSize)
+		throws SchedulerException {
+
+		Message message = new Message();
+
+		message.put(SchedulerEngine.LANGUAGE, language);
+		message.put(SchedulerEngine.SCRIPT, script);
+
+		schedule(
+			trigger, storageType, description,
+			DestinationNames.SCHEDULER_SCRIPTING, message, exceptionsMaxSize);
+	}
+
+	public void auditSchedulerJobs(Message message, TriggerState triggerState)
+		throws SchedulerException {
+
+		if (!PropsValues.AUDIT_MESSAGE_SCHEDULER_JOB) {
+			return;
+		}
+
+		try {
+			AuditMessage auditMessage = new AuditMessage(
+				SchedulerEngine.AUDIT_ACTION, CompanyConstants.SYSTEM, 0,
+				StringPool.BLANK, SchedulerEngine.class.getName(), "0",
+				triggerState.toString(), new Date(),
+				JSONFactoryUtil.createJSONObject(
+					JSONFactoryUtil.serialize(message)));
+
+			auditMessage.setServerName(InetAddressUtil.getLocalHostName());
+			auditMessage.setServerPort(PortalUtil.getPortalPort(false));
+
+			AuditRouterUtil.route(auditMessage);
+		}
+		catch (Exception e) {
+			throw new SchedulerException(e);
+		}
+	}
+
+	public void delete(SchedulerEntry schedulerEntry, StorageType storageType)
+		throws SchedulerException {
+
+		Trigger trigger = schedulerEntry.getTrigger();
+
+		delete(trigger.getJobName(), trigger.getGroupName(), storageType);
+	}
+
+	public void delete(String groupName, StorageType storageType)
+		throws SchedulerException {
+
+		_schedulerEngine.delete(namespaceGroupName(groupName, storageType));
+	}
+
+	public void delete(
+			String jobName, String groupName, StorageType storageType)
+		throws SchedulerException {
+
+		_schedulerEngine.delete(
+			jobName, namespaceGroupName(groupName, storageType));
+	}
+
+	public String getCronText(Calendar calendar, boolean timeZoneSensitive) {
+		return getCronText(
+			null, calendar, timeZoneSensitive, Recurrence.NO_RECURRENCE);
+	}
+
+	public String getCronText(
+		PortletRequest portletRequest, Calendar calendar,
+		boolean timeZoneSensitive, int recurrenceType) {
+
+		Calendar recurrenceCalendar = null;
+
+		if (timeZoneSensitive) {
+			recurrenceCalendar = CalendarFactoryUtil.getCalendar();
+
+			recurrenceCalendar.setTime(calendar.getTime());
+		}
+		else {
+			recurrenceCalendar = (Calendar)calendar.clone();
+		}
+
+		Recurrence recurrence = new Recurrence(
+			recurrenceCalendar, new Duration(1, 0, 0, 0), recurrenceType);
+
+		recurrence.setWeekStart(Calendar.SUNDAY);
+
+		if (recurrenceType == Recurrence.DAILY) {
+			int dailyType = ParamUtil.getInteger(portletRequest, "dailyType");
+
+			if (dailyType == 0) {
+				int dailyInterval = ParamUtil.getInteger(
+					portletRequest, "dailyInterval", 1);
+
+				recurrence.setInterval(dailyInterval);
+			}
+			else {
+				DayAndPosition[] dayPos = {
+					new DayAndPosition(Calendar.MONDAY, 0),
+					new DayAndPosition(Calendar.TUESDAY, 0),
+					new DayAndPosition(Calendar.WEDNESDAY, 0),
+					new DayAndPosition(Calendar.THURSDAY, 0),
+					new DayAndPosition(Calendar.FRIDAY, 0)};
+
+				recurrence.setByDay(dayPos);
+			}
+		}
+		else if (recurrenceType == Recurrence.WEEKLY) {
+			int weeklyInterval = ParamUtil.getInteger(
+				portletRequest, "weeklyInterval", 1);
+
+			recurrence.setInterval(weeklyInterval);
+
+			List<DayAndPosition> dayPos = new ArrayList<DayAndPosition>();
+
+			_addWeeklyDayPos(portletRequest, dayPos, Calendar.SUNDAY);
+			_addWeeklyDayPos(portletRequest, dayPos, Calendar.MONDAY);
+			_addWeeklyDayPos(portletRequest, dayPos, Calendar.TUESDAY);
+			_addWeeklyDayPos(portletRequest, dayPos, Calendar.WEDNESDAY);
+			_addWeeklyDayPos(portletRequest, dayPos, Calendar.THURSDAY);
+			_addWeeklyDayPos(portletRequest, dayPos, Calendar.FRIDAY);
+			_addWeeklyDayPos(portletRequest, dayPos, Calendar.SATURDAY);
+
+			if (dayPos.size() == 0) {
+				dayPos.add(new DayAndPosition(Calendar.MONDAY, 0));
+			}
+
+			recurrence.setByDay(dayPos.toArray(new DayAndPosition[0]));
+		}
+		else if (recurrenceType == Recurrence.MONTHLY) {
+			int monthlyType = ParamUtil.getInteger(
+				portletRequest, "monthlyType");
+
+			if (monthlyType == 0) {
+				int monthlyDay = ParamUtil.getInteger(
+					portletRequest, "monthlyDay0", 1);
+
+				recurrence.setByMonthDay(new int[] {monthlyDay});
+
+				int monthlyInterval = ParamUtil.getInteger(
+					portletRequest, "monthlyInterval0", 1);
+
+				recurrence.setInterval(monthlyInterval);
+			}
+			else {
+				int monthlyPos = ParamUtil.getInteger(
+					portletRequest, "monthlyPos");
+				int monthlyDay = ParamUtil.getInteger(
+					portletRequest, "monthlyDay1");
+
+				DayAndPosition[] dayPos = {
+					new DayAndPosition(monthlyDay, monthlyPos)};
+
+				recurrence.setByDay(dayPos);
+
+				int monthlyInterval = ParamUtil.getInteger(
+					portletRequest, "monthlyInterval1", 1);
+
+				recurrence.setInterval(monthlyInterval);
+			}
+		}
+		else if (recurrenceType == Recurrence.YEARLY) {
+			int yearlyType = ParamUtil.getInteger(portletRequest, "yearlyType");
+
+			if (yearlyType == 0) {
+				int yearlyMonth = ParamUtil.getInteger(
+					portletRequest, "yearlyMonth0");
+				int yearlyDay = ParamUtil.getInteger(
+					portletRequest, "yearlyDay0", 1);
+
+				recurrence.setByMonth(new int[] {yearlyMonth});
+				recurrence.setByMonthDay(new int[] {yearlyDay});
+
+				int yearlyInterval = ParamUtil.getInteger(
+					portletRequest, "yearlyInterval0", 1);
+
+				recurrence.setInterval(yearlyInterval);
+			}
+			else {
+				int yearlyPos = ParamUtil.getInteger(
+					portletRequest, "yearlyPos");
+				int yearlyDay = ParamUtil.getInteger(
+					portletRequest, "yearlyDay1");
+				int yearlyMonth = ParamUtil.getInteger(
+					portletRequest, "yearlyMonth1");
+
+				DayAndPosition[] dayPos = {
+					new DayAndPosition(yearlyDay, yearlyPos)};
+
+				recurrence.setByDay(dayPos);
+
+				recurrence.setByMonth(new int[] {yearlyMonth});
+
+				int yearlyInterval = ParamUtil.getInteger(
+					portletRequest, "yearlyInterval1", 1);
+
+				recurrence.setInterval(yearlyInterval);
+			}
+		}
+
+		return RecurrenceSerializer.toCronText(recurrence);
+	}
+
+	public Date getEndTime(SchedulerResponse schedulerResponse) {
+		Message message = schedulerResponse.getMessage();
+
+		JobState jobState = (JobState)message.get(SchedulerEngine.JOB_STATE);
+
+		TriggerState triggerState = jobState.getTriggerState();
+
+		if (triggerState.equals(TriggerState.NORMAL) ||
+			triggerState.equals(TriggerState.PAUSED)) {
+
+			return (Date)message.get(SchedulerEngine.END_TIME);
+		}
+		else {
+			return jobState.getTriggerDate(SchedulerEngine.END_TIME);
+		}
+	}
+
+	public Date getEndTime(
+			String jobName, String groupName, StorageType storageType)
+		throws SchedulerException {
+
+		SchedulerResponse schedulerResponse = getScheduledJob(
+			jobName, groupName, storageType);
+
+		if (schedulerResponse != null) {
+			return getEndTime(schedulerResponse);
+		}
+
+		return null;
+	}
+
+	public Date getFinalFireTime(SchedulerResponse schedulerResponse) {
+		Message message = schedulerResponse.getMessage();
+
+		JobState jobState = (JobState)message.get(SchedulerEngine.JOB_STATE);
+
+		TriggerState triggerState = jobState.getTriggerState();
+
+		if (triggerState.equals(TriggerState.NORMAL) ||
+			triggerState.equals(TriggerState.PAUSED)) {
+
+			return (Date)message.get(SchedulerEngine.FINAL_FIRE_TIME);
+		}
+		else {
+			return jobState.getTriggerDate(SchedulerEngine.FINAL_FIRE_TIME);
+		}
+	}
+
+	public Date getFinalFireTime(
+			String jobName, String groupName, StorageType storageType)
+		throws SchedulerException {
+
+		SchedulerResponse schedulerResponse = getScheduledJob(
+			jobName, groupName, storageType);
+
+		if (schedulerResponse != null) {
+			return getFinalFireTime(schedulerResponse);
+		}
+
+		return null;
+	}
+
+	public ObjectValuePair<Exception, Date>[] getJobExceptions(
+		SchedulerResponse schedulerResponse) {
+
+		Message message = schedulerResponse.getMessage();
+
+		JobState jobState = (JobState)message.get(SchedulerEngine.JOB_STATE);
+
+		return jobState.getExceptions();
+	}
+
+	public ObjectValuePair<Exception, Date>[] getJobExceptions(
+			String jobName, String groupName, StorageType storageType)
+		throws SchedulerException {
+
+		SchedulerResponse schedulerResponse = getScheduledJob(
+			jobName, groupName, storageType);
+
+		if (schedulerResponse != null) {
+			return getJobExceptions(schedulerResponse);
+		}
+
+		return null;
+	}
+
+	public TriggerState getJobState(SchedulerResponse schedulerResponse) {
+		Message message = schedulerResponse.getMessage();
+
+		JobState jobState = (JobState)message.get(SchedulerEngine.JOB_STATE);
+
+		return jobState.getTriggerState();
+	}
+
+	public TriggerState getJobState(
+			String jobName, String groupName, StorageType storageType)
+		throws SchedulerException {
+
+		SchedulerResponse schedulerResponse = getScheduledJob(
+			jobName, groupName, storageType);
+
+		if (schedulerResponse != null) {
+			return getJobState(schedulerResponse);
+		}
+
+		return null;
+	}
+
+	public Date getNextFireTime(SchedulerResponse schedulerResponse) {
+		Message message = schedulerResponse.getMessage();
+
+		JobState jobState = (JobState)message.get(SchedulerEngine.JOB_STATE);
+
+		TriggerState triggerState = jobState.getTriggerState();
+
+		if (triggerState.equals(TriggerState.NORMAL) ||
+			triggerState.equals(TriggerState.PAUSED)) {
+
+			return (Date)message.get(SchedulerEngine.NEXT_FIRE_TIME);
+		}
+		else {
+			return jobState.getTriggerDate(SchedulerEngine.NEXT_FIRE_TIME);
+		}
+	}
+
+	public Date getNextFireTime(
+			String jobName, String groupName, StorageType storageType)
+		throws SchedulerException {
+
+		SchedulerResponse schedulerResponse = getScheduledJob(
+			jobName, groupName, storageType);
+
+		if (schedulerResponse != null) {
+			return getNextFireTime(schedulerResponse);
+		}
+
+		return null;
+	}
+
+	public Date getPreviousFireTime(SchedulerResponse schedulerResponse) {
+		Message message = schedulerResponse.getMessage();
+
+		JobState jobState = (JobState)message.get(SchedulerEngine.JOB_STATE);
+
+		TriggerState triggerState = jobState.getTriggerState();
+
+		if (triggerState.equals(TriggerState.NORMAL) ||
+			triggerState.equals(TriggerState.PAUSED)) {
+
+			return (Date)message.get(SchedulerEngine.PREVIOUS_FIRE_TIME);
+		}
+		else {
+			return jobState.getTriggerDate(SchedulerEngine.PREVIOUS_FIRE_TIME);
+		}
+	}
+
+	public Date getPreviousFireTime(
+			String jobName, String groupName, StorageType storageType)
+		throws SchedulerException {
+
+		SchedulerResponse schedulerResponse = getScheduledJob(
+			jobName, groupName, storageType);
+
+		if (schedulerResponse != null) {
+			return getPreviousFireTime(schedulerResponse);
+		}
+
+		return null;
+	}
+
+	public SchedulerResponse getScheduledJob(
+			String jobName, String groupName, StorageType storageType)
+		throws SchedulerException {
+
+		return _schedulerEngine.getScheduledJob(
+			jobName, namespaceGroupName(groupName, storageType));
+	}
+
+	public List<SchedulerResponse> getScheduledJobs()
+		throws SchedulerException {
+
+		return _schedulerEngine.getScheduledJobs();
+	}
+
+	public List<SchedulerResponse> getScheduledJobs(StorageType storageType)
+		throws SchedulerException {
+
+		List<SchedulerResponse> schedulerResponses =
+			new ArrayList<SchedulerResponse>();
+
+		for (SchedulerResponse schedulerResponse :
+				_schedulerEngine.getScheduledJobs()) {
+
+			if (storageType.equals(schedulerResponse.getStorageType())) {
+				schedulerResponses.add(schedulerResponse);
+			}
+		}
+
+		return schedulerResponses;
+	}
+
+	public List<SchedulerResponse> getScheduledJobs(
+			String groupName, StorageType storageType)
+		throws SchedulerException {
+
+		return _schedulerEngine.getScheduledJobs(
+			namespaceGroupName(groupName, storageType));
+	}
+
+	public Date getStartTime(SchedulerResponse schedulerResponse) {
+		Message message = schedulerResponse.getMessage();
+
+		JobState jobState = (JobState)message.get(SchedulerEngine.JOB_STATE);
+
+		TriggerState triggerState = jobState.getTriggerState();
+
+		if (triggerState.equals(TriggerState.NORMAL) ||
+			triggerState.equals(TriggerState.PAUSED)) {
+
+			return (Date)message.get(SchedulerEngine.START_TIME);
+		}
+		else {
+			return jobState.getTriggerDate(SchedulerEngine.START_TIME);
+		}
+	}
+
+	public Date getStartTime(
+			String jobName, String groupName, StorageType storageType)
+		throws SchedulerException {
+
+		SchedulerResponse schedulerResponse = getScheduledJob(
+			jobName, groupName, storageType);
+
+		if (schedulerResponse != null) {
+			return getStartTime(schedulerResponse);
+		}
+
+		return null;
+	}
+
+	public void initialize() throws SchedulerException {
+		if (_schedulerEngineClusterManager != null) {
+			_schedulerEngineClusterManager.initialize();
+		}
+
+		SchedulerLifecycle schedulerLifecycle = new SchedulerLifecycle();
+
+		schedulerLifecycle.registerPortalLifecycle(PortalLifecycle.METHOD_INIT);
+	}
+
+	public String namespaceGroupName(
+		String groupName, StorageType storageType) {
+
+		return storageType.toString().concat(StringPool.POUND).concat(
+			groupName);
+	}
+
+	public void pause(String groupName, StorageType storageType)
+		throws SchedulerException {
+
+		_schedulerEngine.pause(namespaceGroupName(groupName, storageType));
+	}
+
+	public void pause(String jobName, String groupName, StorageType storageType)
+		throws SchedulerException {
+
+		_schedulerEngine.pause(
+			jobName, namespaceGroupName(groupName, storageType));
+	}
+
+	public void resume(String groupName, StorageType storageType)
+		throws SchedulerException {
+
+		_schedulerEngine.resume(namespaceGroupName(groupName, storageType));
+	}
+
+	public void resume(
+			String jobName, String groupName, StorageType storageType)
+		throws SchedulerException {
+
+		_schedulerEngine.resume(
+			jobName, namespaceGroupName(groupName, storageType));
+	}
+
+	public void schedule(
+			SchedulerEntry schedulerEntry, StorageType storageType,
+			String portletId, int exceptionsMaxSize)
+		throws SchedulerException {
+
+		Message message = new Message();
+
+		message.put(
+			SchedulerEngine.CONTEXT_PATH, schedulerEntry.getContextPath());
+		message.put(
+			SchedulerEngine.MESSAGE_LISTENER_CLASS_NAME,
+			schedulerEntry.getEventListenerClass());
+		message.put(SchedulerEngine.PORTLET_ID, portletId);
+
+		schedule(
+			schedulerEntry.getTrigger(), storageType,
+			schedulerEntry.getDescription(),
+			DestinationNames.SCHEDULER_DISPATCH, message, exceptionsMaxSize);
+	}
+
+	public void schedule(
+			Trigger trigger, StorageType storageType, String description,
+			String destinationName, Message message, int exceptionsMaxSize)
+		throws SchedulerException {
+
+		if (message == null) {
+			message = new Message();
+		}
+
+		message.put(SchedulerEngine.EXCEPTIONS_MAX_SIZE, exceptionsMaxSize);
+
+		trigger = TriggerFactoryUtil.buildTrigger(
+			trigger.getTriggerType(), trigger.getJobName(),
+			namespaceGroupName(trigger.getGroupName(), storageType),
+			trigger.getStartDate(), trigger.getEndDate(),
+			trigger.getTriggerContent());
+
+		_schedulerEngine.schedule(
+			trigger, description, destinationName, message);
+	}
+
+	public void schedule(
+			Trigger trigger, StorageType storageType, String description,
+			String destinationName, Object payload, int exceptionsMaxSize)
+		throws SchedulerException {
+
+		Message message = new Message();
+
+		message.setPayload(payload);
+
+		schedule(
+			trigger, storageType, description, destinationName, message,
+			exceptionsMaxSize);
+	}
+
+	public void setSchedulerEngine(SchedulerEngine schedulerEngine) {
+		_schedulerEngine = schedulerEngine;
+
+		if (schedulerEngine instanceof SchedulerEngineClusterManager) {
+			_schedulerEngineClusterManager =
+				(SchedulerEngineClusterManager)schedulerEngine;
+		}
+	}
+
+	public void shutdown() throws SchedulerException {
+		_schedulerEngine.shutdown();
+	}
+
+	public void start() throws SchedulerException {
+		_schedulerEngine.start();
+	}
+
+	public void suppressError(
+			String jobName, String groupName, StorageType storageType)
+		throws SchedulerException {
+
+		_schedulerEngine.suppressError(
+			jobName, namespaceGroupName(groupName, storageType));
+	}
+
+	public void unschedule(
+			SchedulerEntry schedulerEntry, StorageType storageType)
+		throws SchedulerException {
+
+		Trigger trigger = schedulerEntry.getTrigger();
+
+		unschedule(trigger.getJobName(), trigger.getGroupName(), storageType);
+	}
+
+	public void unschedule(String groupName, StorageType storageType)
+		throws SchedulerException {
+
+		_schedulerEngine.unschedule(namespaceGroupName(groupName, storageType));
+	}
+
+	public void unschedule(
+			String jobName, String groupName, StorageType storageType)
+		throws SchedulerException {
+
+		_schedulerEngine.unschedule(
+			jobName, namespaceGroupName(groupName, storageType));
+	}
+
+	public void update(
+			String jobName, String groupName, StorageType storageType,
+			String description, String language, String script,
+			int exceptionsMaxSize)
+		throws SchedulerException {
+
+		SchedulerResponse schedulerResponse = getScheduledJob(
+			jobName, groupName, storageType);
+
+		if (schedulerResponse == null) {
+			return;
+		}
+
+		Trigger trigger = schedulerResponse.getTrigger();
+
+		if (trigger == null) {
+			return;
+		}
+
+		Message message = schedulerResponse.getMessage();
+
+		if (message == null) {
+			return;
+		}
+
+		addScriptingJob(
+			trigger, storageType, description, language, script,
+			exceptionsMaxSize);
+	}
+
+	public void update(Trigger trigger, StorageType storageType)
+		throws SchedulerException {
+
+		trigger = TriggerFactoryUtil.buildTrigger(
+			trigger.getTriggerType(), trigger.getJobName(),
+			namespaceGroupName(trigger.getGroupName(), storageType),
+			trigger.getStartDate(), trigger.getEndDate(),
+			trigger.getTriggerContent());
+
+		_schedulerEngine.update(trigger);
+	}
+
+	public void updateMemorySchedulerClusterMaster() throws SchedulerException {
+		if (_schedulerEngineClusterManager == null) {
+			_log.error(
+				"Unable to update memory scheduler cluster master because " +
+					"the portal is not using a clustered scheduler engine");
+
+			return;
+		}
+
+		_schedulerEngineClusterManager.updateMemorySchedulerClusterMaster();
+	}
+
+	private void _addWeeklyDayPos(
+		PortletRequest portletRequest, List<DayAndPosition> list, int day) {
+
+		if (ParamUtil.getBoolean(portletRequest, "weeklyDayPos" + day)) {
+			list.add(new DayAndPosition(day, 0));
+		}
+	}
+
+	private static Log _log = LogFactoryUtil.getLog(
+		SchedulerEngineHelperImpl.class);
+
+	private SchedulerEngine _schedulerEngine;
+	private SchedulerEngineClusterManager _schedulerEngineClusterManager;
+
+}
