+++ portal-impl/src/portal.properties
@@ -5505,6 +5505,7 @@
 ##
 
     audit.message.com.liferay.portal.model.Layout.VIEW=false
+    audit.message.scheduler.job=false
 
 ##
 ## Buffered Increment
