+++ portal-service/src/com/liferay/portal/kernel/util/PropsKeys.java
@@ -119,6 +119,8 @@ public interface PropsKeys {
 
 	public static final String AUDIT_MESSAGE_COM_LIFERAY_PORTAL_MODEL_LAYOUT_VIEW = "audit.message.com.liferay.portal.model.Layout.VIEW";
 
+	public static final String AUDIT_MESSAGE_SCHEDULER_JOB = "audit.message.scheduler.job";
+
 	public static final String AUTH_FAILURE = "auth.failure";
 
 	public static final String AUTH_FORWARD_BY_LAST_PATH = "auth.forward.by.last.path";
